Fix compiler warnings across codebase
- Crypto.cpp: replace deprecated MD5() with EVP digest API (OpenSSL 3.0)
- Nvidia.cpp: add static_cast<int> for int64_t narrowing conversions
- CPU.cpp: fix sign-compare, catch-by-value, and narrowing warnings
- Tree.hpp: use const reference in range-for loop to avoid copying
- DragChartView.cpp: replace deprecated QPalette::Background with Window
- DeviceTreeView.cpp: fix sign-compare with size_t loop variable
- EnumEditor.cpp: replace deprecated QVector::fromStdVector, fix sign-compare
- tuxclockerd/main.cpp: suppress unused variable warning with Q_UNUSED

+++ src/tuxclocker-qt/widgets/DeviceTreeView.cpp
@@ -140,7 +140,7 @@ bool assignableInfoEquals(AssignableInfo a, AssignableInfo b) {
 		if (enumsA.size() != enumsB.size())
 			return false;
 
-		for (int i = 0; i < enumsA.size(); i++) {
+		for (size_t i = 0; i < enumsA.size(); i++) {
 			if (enumsA[i].key != enumsB[i].key || enumsA[i].name != enumsB[i].name)
 				return false;
 		}
